+++ app/Http/Controllers/User/ArticleController.php
@@ -6,6 +6,8 @@ use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
 use GirdPlugins\Base\ArticleFunc;
+use GirdPlugins\Base\LogFunc;
+use GirdPlugins\Base\UserPowerFunc;
 class ArticleController extends Controller 
 {
 
@@ -13,7 +15,6 @@ class ArticleController extends Controller
 
     public function sArticle(ArticleFunc $atcFunc)
     {
-
         $articleData = $atcFunc->getUserArticle(session("user.user_id"));
         $inputData["articleData"] = $articleData;
         session(["nowPage"=>"/user_sArticle"]);
@@ -32,29 +33,52 @@ class ArticleController extends Controller
         return view("User.Article.aArticle",$inputData);
         
     }
-    public function _aArticle(ArticleFunc $atcFunc,  BaseFunc $baseFunc)
-    {        
-        
+    public function _aArticle(UserPowerFunc $userPowerFunc,LogFunc $logFunc,ArticleFunc $atcFunc,  BaseFunc $baseFunc)
+    {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $articleData = Request::only("article_title","article_intro","article_class","article_sort","article_detail");
-        
+        DB::beginTransaction();
         if(true == $atcFunc->addArticle($articleData))
         {
+            $log_array['log_level']=0;
+            $log_array['log_title']="添加操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."添加了一篇文章";
+            $log_array['log_data']="添加";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
             return response()->json(['status' => true, 'message' => '<p class="text-success">修改成功，即将跳转</p>']);
         }
         else
         {
-
-           
+          
            return response()->json(['status' => true, 'message' => '<p class="text-success">修改成功，即将跳转</p>']);
-
            
-        }
-        
+        }        
     }
-    public function dArticle($articleId,ArticleFunc $atcFunc,  BaseFunc $baseFunc)
+    public function dArticle(UserPowerFunc $userPowerFunc,LogFunc $logFunc,$articleId,ArticleFunc $atcFunc,  BaseFunc $baseFunc)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
+        DB::beginTransaction();
         if($atcFunc->deleteUserArticle(session("user.user_id"),$articleId))
         {
+            $log_array['log_level']=0;
+            $log_array['log_title']="删除操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."删除了一篇文章";
+            $log_array['log_data']="删除";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
             $baseFunc->setRedirectMessage(true, "成功删除", NULL,"/user_sArticle");
         }
         else
@@ -78,14 +102,28 @@ class ArticleController extends Controller
         $data = Request::only("article_id");
         echo $atcFunc->getArticleDetail( $data["article_id"])->article_detail;
     }
-    public function _uArticle() 
+    public function _uArticle(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc) 
    {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $articleData = Request::only("article_title","article_intro","article_class","article_sort","article_detail");
         $articleId = Request::input("article_id");
+        DB::beginTransaction();
         if(DB::table("base_article")->where("article_id","=",$articleId)
                 ->where("article_user","=",session("user.user_id"))
                 ->update($articleData))
         {
+            $log_array['log_level']=0;
+            $log_array['log_title']="修改操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."修改了一篇文章";
+            $log_array['log_data']="修改";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
             return response()->json(['status' => true, 'message' => '<p class="text-success">修改成功，即将跳转</p>']);
         }
         else
@@ -99,16 +137,30 @@ class ArticleController extends Controller
         session(["nowPage"=>"/user_sSubject"]);
         return view("User.Article.sSubject",$data);
     }
-    public function aSubject(BaseFunc $base)
+    public function aSubject(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $base)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $subjectData = Request::only("subject_name","subject_intro"); 
         $subjectData['subject_create_date']=date("Y-m-d H:i:s");
         $subjectData['subject_update_date']=date("Y-m-d H:i:s");
         $subjectData['subject_user']=session("user.user_id");
+        DB::beginTransaction();
         if(DB::table("base_article_subject")->insert($subjectData))
         {
             //添加成功，提示跳转
             $base->setRedirectMessage(true, "添加专题成功！", null, null);
+            $log_array['log_level']=0;
+            $log_array['log_title']="添加操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."添加了一个专题";
+            $log_array['log_data']="添加";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
             return redirect()->back();
         }
         else
@@ -120,12 +172,19 @@ class ArticleController extends Controller
         //dump($subject_add_data);
     }
 
-    public function uSubject(BaseFunc $baseFunc)
+    public function uSubject(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $subjectData = Request::only("subject_name","subject_intro");
         $subjectId = Request::input("subject_id");
         $subjectData['subject_name']=$subjectData['subject_name'];
         $subjectData['subject_update_date']=  date("Y-m-d H:i:s");
+        DB::beginTransaction();
         if(DB::table("base_article_subject")
                 ->where("subject_id","=",$subjectId)
                 ->where("subject_user","=",session("user.user_id"))
@@ -133,6 +192,13 @@ class ArticleController extends Controller
         {
             //修改成功，提示跳转
             $baseFunc->setRedirectMessage(true, "修改专题成功！", null, null);
+            $log_array['log_level']=0;
+            $log_array['log_title']="修改操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."修改了一个专题";
+            $log_array['log_data']="修改";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
             return redirect()->back();
         }
         else
@@ -144,14 +210,27 @@ class ArticleController extends Controller
         //dump($sunject_update_data);
     }
     
-    public function dSubject(BaseFunc $baseFunc,$subject_id)
+    public function dSubject(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc,$subject_id)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         //删除base_article_subject表的指定专题
         DB::table("base_article_subject")
                 ->where("subject_id","=","$subject_id")
                  ->where("subject_user","=",session("user.user_id"))
                 ->delete();
         $baseFunc->setRedirectMessage(true, "删除专题成功！", null, null);
+        $log_array['log_level']=0;
+        $log_array['log_title']="删除操作";
+        $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."删除了一个专题";
+        $log_array['log_data']="删除";
+        $log_array['log_user']=session("user.user_id");
+        $logFunc->addLog($log_array);
+        DB::commit();
         return redirect()->back();
     }
     
@@ -176,30 +255,58 @@ class ArticleController extends Controller
         return view("User.Article.moreSubject",$inputData);
     }
     
-    public function addArticleToSubject(BaseFunc $baseFunc)
+    public function addArticleToSubject(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $postData = Request::only("subject_id", "article_id_array");
         if($postData['article_id_array'] == "")
         {
           $baseFunc->setRedirectMessage(false, "没有选择文章！", NULL, NULL);
           return redirect()->back();
         }
+        DB::beginTransaction();
         foreach ($postData["article_id_array"] as $data) 
         {
           DB::table("base_article_re_subject")->insert(["relation_article" => $data, "relation_subject" => $postData["subject_id"]]);
         }
         $baseFunc->setRedirectMessage(true, "添加文章成功！", NULL, NULL);
+        $log_array['log_level']=0;
+        $log_array['log_title']="添加操作";
+        $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."添加了一篇文章到一个专题";
+        $log_array['log_data']="添加";
+        $log_array['log_user']=session("user.user_id");
+        $logFunc->addLog($log_array);
+        DB::commit();
         return redirect()->back();
     }
 
     
-    public function removeArticleToSubject(BaseFunc $baseFunc,$subject_id,$article_id)
+    public function removeArticleToSubject(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc,$subject_id,$article_id)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
+        DB::beginTransaction();
         DB::table("base_article_re_subject")
                 ->where("relation_subject","=",$subject_id)
                 ->where("relation_article","=",$article_id)
                 ->update(["relation_subject"=>null]);
         $baseFunc->setRedirectMessage(true, "移除文章成功！", null, null);
+        $log_array['log_level']=0;
+            $log_array['log_title']="移除操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."从一个专题溢出了一篇文章";
+            $log_array['log_data']="移除";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
         return redirect()->back();
     }
 
@@ -210,31 +317,72 @@ class ArticleController extends Controller
         return view("User.Article.sLabel",$data);
     }
     
-    public function aLabel(BaseFunc $baseFunc)
+    public function aLabel(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $input_data = Request::only("label_name");
         $input_data['label_create_date']=date("Y-m-d H:i:s");
         $input_data['label_update_date']= date("Y-m-d H:i:s");
+        DB::beginTransaction();
         DB::table("base_article_label")->insert($input_data);
         $baseFunc->setRedirectMessage(true, "添加标签成功！", null, null);
+        $log_array['log_level']=0;
+            $log_array['log_title']="添加操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."添加了一个标签";
+            $log_array['log_data']="添加";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
         return redirect()->back();
     }
     
-    public function uLabel(BaseFunc $baseFunc)
+    public function uLabel(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         $input_data = Request::only("label_id","label_name","label_update_date");
         $input_data['label_update_date']= date("Y-m-d H:i:s");
+        DB::beginTransaction();
         DB::table("base_article_label")
                 ->where("label_id","=",$input_data['label_id'])
                 ->update($input_data);
         $baseFunc->setRedirectMessage(true, "标签修改成功",null,null);
+        $log_array['log_level']=0;
+            $log_array['log_title']="修改操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."修改了一个标签";
+            $log_array['log_data']="修改";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
         return redirect()->back();    
     }
     
-    public function dLabel(BaseFunc $baseFunc,$label_id)
+    public function dLabel(UserPowerFunc $userPowerFunc,LogFunc $logFunc,BaseFunc $baseFunc,$label_id)
     {
+        $powerId=10;
+        if($userPowerFunc->checkUserPower($powerId))
+        {
+            $baseFunc->setRedirectMessage(false, "你没有权限进行此操作，请联系超级管理员", NULL, NULL);
+            return redirect()->back();
+        }
         DB::table("base_article_label")->where("label_id","=",$label_id)->delete();
         $baseFunc->setRedirectMessage(true, "删除专题成功！", null, null);
+        $log_array['log_level']=0;
+            $log_array['log_title']="添加操作";
+            $log_array['log_detail']=date("Y-m-d H:i:s").session('user.user_nickname')."添加了一个标签";
+            $log_array['log_data']="添加";
+            $log_array['log_user']=session("user.user_id");
+            $logFunc->addLog($log_array);
+            DB::commit();
         return redirect()->back();
     }
 
@@ -292,12 +440,49 @@ class ArticleController extends Controller
     
     public function sCollect()
     {
-        session(["nowPage"=>"/user_sCollect"]);
-        $data=DB::table('base_article_collect')->get();
+        session(["nowPage"=>"/user_sCollect"]); 
+        $data["collectData"]=DB::table('base_article_collect')
+                ->leftJoin("base_collect_relation","relation_id","=","collect_id")
+                ->leftJoin("base_user","user_id","=","collect_id")
+                ->leftJoin("base_article","article_id","=","collect_article_id")
+                ->get();
+        
         return view("User.Article.sCollect",$data);
     }
     
+    public function aCollect(LogFunc $logFunc,BaseFunc $baseFunc)
+    {
+        $input_data=Request::only("collect_name");
+        $input_data['collect_create_date']=date("Y-m-d H:i:s");
+        $input_data['collect_update_date']= date("Y-m-d H:i:s");
+        $input_data['collect_user']=  session("user.user_id");
+        $input_data['collect_root']=0;
+        $input_data['collect_folder']=1;
+        DB::table("base_article_collect")->insert($input_data);
+        $baseFunc->setRedirectMessage(true, "添加收藏夹成功！", null, null);
+        return redirect()->back();
+    }
     
+    public function moreCollect($collect_id)
+    { 
+        $moreCollect=DB::table('base_collect_relation')
+                ->leftJoin("base_article_collect","collect_id","=","relation_child")
+                ->leftJoin("base_article","article_id","=","collect_article_id")
+                ->leftJoin("base_user","user_id","=","article_user")
+                ->where("relation_parent","=","$collect_id")
+                ->get();
+        $data["moreCollect"]=$moreCollect;
+        $data["nowCollect"]=DB::table('base_article_collect')
+                ->where("collect_id","=","$collect_id")
+                ->get();    
+        return view("User.Article.moreCollect",$data);
+    }
+
+
+
+
+
+
 
     //处理文章评论的函数（在这里评论提交的表单是默认为是头条评论，也就是relation_parent=0）
